+++ flutter_local_notifications/test/android_flutter_local_notifications_test.dart
@@ -1954,6 +1954,7 @@ void main() {
               'notification body',
               repeatInterval,
               const NotificationDetails(android: androidNotificationDetails),
+              androidScheduleMode: AndroidScheduleMode.exact,
             );
 
             expect(